feat(shs-5963): move vertical tabs js to hs_layouts and attach to it, remove index.js in humsci_basic
