+++ docroot/modules/humsci/hs_layouts/hs_layouts.layouts.yml
@@ -20,6 +20,7 @@ three_column:
   template: layouts/three-column/three-column
   default_region: main
   class: '\Drupal\hs_layouts\Plugin\Layout\HumsciLayout'
+  library: hs_layouts/verticaltabs
   icon_map:
     - [sidebar_first, main, sidebar_second]
   regions:
