+++ docroot/modules/humsci/hs_layouts/hs_layouts.libraries.yml
@@ -2,10 +2,20 @@ three_column_w_image:
   css:
     layout:
       css/three-column-w-image.css: {}
+  js:
+    js/vertical-tabs.js: {}
+  dependencies:
+    - core/drupal
+
 news_style:
   css:
     layout:
       css/news-style.css: {}
+  js:
+    js/vertical-tabs.js: {}
+  dependencies:
+    - core/drupal
+
 layout_builder_admin:
   css:
     theme:
@@ -19,4 +29,10 @@ images_loaded:
     //imagesloaded.desandro.com/imagesloaded.pkgd.min.js: {type: external, minified: true }
   dependencies:
     - core/jquery
+
+verticaltabs:
+  js:
+    js/vertical-tabs.js: {}
+  dependencies:
+    - core/drupal
