+++ docroot/modules/humsci/hs_layouts/js/vertical-tabs.js
@@ -1,0 +1,13 @@
+function closeDetails() {
+// Close Revision Information Details element in Layout Builder by default.
+  if (document.querySelector('.layout-builder-form')) {
+    const details = document.querySelector('.layout-builder-form details');
+    if (details) {
+      details.removeAttribute('open');
+    }
+  }
+}
+
+document.addEventListener('DOMContentLoaded', () => {
+  closeDetails(document);
+});
