--- docroot/themes/humsci/humsci_basic/src/js/shared/index.js
@@ -1,20 +1,0 @@
-import './carousel-slides/carousel-slides-height';
-import './tables/wrap';
-import './tables/scope';
-import './tables/table-pattern';
-import './navigation/index';
-import './equal-height-grid/index';
-import './media/video-with-caption';
-import './animation/page-scroll';
-import './timeline/expand-collapse-timeline';
-import './accordion/accordion-toggle-all';
-import './linked-cards/linked-cards';
-import './photo-album/colorbox';
-import './editoria11y/editoria11y';
-import './vertical-tabs/vertical-tabs';
-import './addtocal/addtocal';
-import './prefered-reduced-motion/prefered-reduced-motion';
-import './main-content-fallback/main-content-fallback';
-
-// PoC for new megamenu.
-import './megamenu/index';
